bryan-dokan_adapter/client: Fixed Visual Studio compile warnings (mainly signed/unsigned issues). Ignored warnings when necessary.
git-svn-id: https://xtreemfs.googlecode.com/svn/branches/bryan-dokan_adapter@3347 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ cpp/include/libxtreemfs/async_write_handler.h
@@ -200,7 +200,7 @@ class AsyncWriteHandler
   std::list<AsyncWriteBuffer*> writes_in_flight_;
 
   /** Number of pending bytes. */
-  int pending_bytes_;
+  size_t pending_bytes_;
 
   /** Number of pending write requests
    *  NOTE: this does not equal writes_in_flight_.size(), since it also contains
@@ -263,10 +263,10 @@ class AsyncWriteHandler
                                //            when async writes support inerrupts
 
   /** Maximum number in bytes which may be pending. */
-  const int max_writeahead_;
+  const size_t max_writeahead_;
 
   /** Maximum number of pending write requests. */
-  const int max_writeahead_requests_;
+  const size_t max_writeahead_requests_;
 
   /** Maximum number of attempts a write will be tried. */
   const int max_write_tries_;
